Removed empty.txt from all 'empty' directories - i.e. directories that are provided as empty by the application, and into which user files will go.
I have instead added .gitignore files to these directories that will ignore everything in them, so that we reduce the risk of accidentally committing user files into the repository, and to make it a lot easier to see the legitimate changes that have been made locally.  The empty.txt files were originally required because Git doesn't allow empty directories, but now we have .gitignore files in them they are no longer empty, so the original file is no longer required.