Fix error when displaying Milestone details
Code would result in error when canEdit=False, as getMilestoneID was being referenced instead of getMilestone.

Also, a lang() construct was used around the Milestone's name. As language strings can't be set by the user (and the Milestone name when elsewhere isn't treated like this) I replaced this with clean().